# marketplace-app/src/main/resources/openapi/marketplace-openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 0.1.0
  description: |
    Complete OpenAPI documentation for the Marketplace backend.
    All business endpoints are prefixed with /api/v1.
    Security: endpoints require bearerAuth (JWT) unless noted otherwise.
    Admin endpoints require the ADMIN role.
servers:
  - url: /
paths:
...
  /api/v1/payments/intents:
    post:
      summary: Create a payment intent
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /api/v1/payments/intents/{id}:
    get:
      summary: Get payment intent by ID
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment intent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /api/v1/payments/intents/{id}/process:
    post:
      summary: Process a payment intent
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment intent processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /api/v1/payments/intents/{id}/confirm:
    post:
      summary: Confirm a payment intent with external ID
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmIntentRequest'
      responses:
        '200':
          description: Payment intent confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /api/v1/payments/intents/{id}/cancel:
    post:
      summary: Cancel a payment intent
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment intent cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /api/v1/payments/{paymentId}/refund:
    post:
      summary: Refund a payment
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /api/v1/payments/webhooks/{provider}:
    post:
      summary: Receive payment provider webhook event
      tags: [Payments]
      security: [{ webhookSignature: [] }]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook accepted

  # ─── Reviews Module ─────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  schemas:
...
    CreateIntentRequest:
      type: object
      required: [bookingId]
      properties:
        bookingId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
    ConfirmIntentRequest:
      type: object
      required: [externalId]
      properties:
        externalId:
          type: string
    PaymentIntentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        amountCents:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amountCents:
          type: integer
          format: int64
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...